Address comments on PR.
